M#-: Fix compilation warnings
(cherry picked from commit 65cedcc2114e640856b59c7c0dcf90b7429422d4)

+++ src/raft/FedReplicaManager.cc
@@ -310,8 +310,8 @@ int FedReplicaManager::get_next_record(int zone_id, std::string& zedp,
         return -2;
     }
 
-    int prev_index = logdb->previous_federated(zs->next);
-	
+    uint64_t prev_index = logdb->previous_federated(zs->next);
+
     if ( prev_index == UINT64_MAX )
     {
         std::ostringstream oss;
@@ -319,10 +319,10 @@ int FedReplicaManager::get_next_record(int zone_id, std::string& zedp,
         oss << "Missing federation record previous to: " << zs->next;
 
         error = oss.str();
-    
+
         pthread_mutex_unlock(&mutex);
 
-	return -1;
+        return -1;
     }
 
     int rc = logdb->get_log_record(zs->next, prev_index, lr);
